First commit of the Snipes NoIRCMan console. If you are testing this and are using the current Snipes (and it's snipes.properties), your username is example and your password is password. If you want to change this, it's in the noircusers/noircpasses properties in snipes.properties.
The only commands in the Snipes console at the moment are: hw (prompts the server to send us back "hello world! :)"); PROTORAW <line> (sends a raw protocol command, the only at the moment is SAYHELLO.); and recv [number of lines (receives the specified amount of lines from the server, or one if not specified). The basics are finished. Oh, also, I lied before. There is one command that is used for both server->client and client<-server (HELLO, with no parameters when sent by a server, with the username and password as arguments on the client); one for client->server only (SAYHELLO, prints back hello world! :)); and three for server->client only (WELCOME, signifies that the sent password/username combination was correct; GOODBYE, signifies that the sent password/username combination was incorrect; and PRNT, which prints a line to the clients standard output).

Parameters are received by receiving a amount of TCP packets equal to the amount of arguments from the stream and sent by sending a amount of packets equal to the amount of arguments to the stream.