First commit in a while. Just did a bit of stress testing on NoIRCMan and fixed the module load command in CoreUtils to check if modules are already loaded.
In detail, there is now a "FLAG" command that is not implemented by the console. It will set a option for what will be sent to the console. The only option at the moment is IRCR, which if set will send all lines from the IRC server to us. I need to fix something in the bot framework to allow for always getting the raw line. Also added a nick specification to snipes.properties.