A lot of file changes, not so many new features.
Made it so all modules can load other modules; added module dependancies (it's not that good at the moment); implemented ModuleReturn (anything other than ERROR means NORMAL, so null is fine); added a IRCUtils class you can optionally load to part and join channels and test the new channels API.

The one BIG feature in this commit is the beginnings of a sort of IRC bouncer. What this bouncer will do is allow you to connect to the bot over port 6667 and basically act as the bot. You will be able to use your IRC client to send commands transparently through the bot. It's not done yet, and all it does is forward all output to a "BotClient" pseudo-client which can only say "Hello world!" at the moment, but soon it will be functioning once authentication works and I learn a little more about how bouncers work. It uses reflection to load pseudo-clients from a enumeration, and then executes it's isLineTo method to see if it is the bot that command is directed to. If all bots say that a line does not belong to them, it will be sent to the last bot (IRCClient) which says that every line is it's and sends them all to the IRC server.

Well, that's me for the night. Sleep time! :P